+++ startup/gui/menus.py
@@ -528,6 +528,7 @@ nodeMenu.append( "/Utility/Context Variables", Gaffer.ContextVariables, searchTe
 nodeMenu.append( "/Utility/Delete Context Variables", Gaffer.DeleteContextVariables, searchText = "DeleteContextVariables" )
 nodeMenu.append( "/Utility/Time Warp", Gaffer.TimeWarp, searchText = "TimeWarp" )
 nodeMenu.append( "/Utility/Spreadsheet", Gaffer.Spreadsheet )
+nodeMenu.append( "/Utility/Context Query", Gaffer.ContextQuery, searchText = "ContextQuery" )
 
 ## Miscellaneous UI
 ###########################################################################